Introduce Field for use with Annotated
We introduce a Field annotation type, to be used with typing.Annotated,
as an alternative way to declare field information of an attr class.
Only a single Field annotation may be supplied in Annotated. A Field
value holds pretty much the same thing as attr.field() except for the
'default' which is expected to be set on the right-hand side.

Validators and converters would be supplied as dedicated annotations, as
argument of typing.Annotated. The Converter class is supported as is but
since there is no Validator class, we'll introduce one next. Several
converters may be specified. On the other hand, if no Field annotation
is present, it is forbidden to use Converter (or Validator) annotations.

As illustrated in added typing example, the 'factory' option is not
handled correctly. This probably needs the mypy.attrs plugin to be
extended.